+++ JoGo/Assets/Scenes/Shop/Prefabs/BuyPopUp.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5e87d259b66c58e41bc88022fee964a2
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
