+++ JoGo/Assets/Scenes/Shop/Prefabs/BuyText.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3abed66c7b4660b42b0e841f886e659b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
