+++ JoGo/Assets/Scenes/Shop/Prefabs/CanCelText.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b3476e8f02bd4b345af2eca1ae818972
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
