+++ JoGo/Assets/Scenes/Shop/Prefabs/IncreaseButton.prefab
@@ -118,12 +118,12 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 5532623209557413178}
-        m_MethodName: 
+      - m_Target: {fileID: 0}
+        m_MethodName: IncreaseOnClick
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: 
+          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
           m_IntArgument: 0
           m_FloatArgument: 0
           m_StringArgument:
