+++ JoGo/Assets/Scenes/Shop/Prefabs/InfoText.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e499d1d824b0cc7479865e1f3ddc6511
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
