+++ JoGo/Assets/Scenes/Shop/Prefabs/ItemName.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3abf905e119139640a1fa7eefb31a5bf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
