--- JoGo/Assets/Scenes/Shop/Scripts/ItemInfo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 641687a1bc58aa34d81b7a3472a51e8f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
