--- JoGo/Assets/Scenes/Shop/Scripts/ItemObject.cs
@@ -1,28 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class ItemObject : MonoBehaviour
-{
-    public Button Item;
-    public Image Face;
-    public GameObject itemInfo;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void OnClick()
-    {
-        itemInfo.SetActive(true);
-    }
-}
