--- JoGo/Assets/Scenes/Shop/Scripts/ItemObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1d509a9a7f1a21749a828825db782ec4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
