+++ JoGo/Assets/Scenes/Shop/Scripts/ShopUIMgr.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 31f14277481d30c4ab40781db4b95306
+guid: 6910b8e9756fac146a81cebed5e04bf6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
